requirements.txt
openai
flask
gunicorn
flask-cors
python-dotenv
